chkPassword 메서드 추가 UserPrincipal 전달받아 오는 과정 추가

+++ src/main/kotlin/com/b2/backoffice/infra/security/UserPrincipal.kt
@@ -4,11 +4,11 @@ import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 
 data class UserPrincipal(
-    val id : Long,
+    val id : Int,
     val email : String,
     val authoricies : Collection<GrantedAuthority>
 ){
-    constructor(id:Long, email:String, role:Set<String>):this(
+    constructor(id:Int, email:String, role:Set<String>):this(
         id,
         email,
         role.map{ SimpleGrantedAuthority("ROLE_$it") }
